various qmd fixes
